Coding journey
Ok, version 0.2 has been successfully released into GIPSC's servers. It
was somehow a struggle running this application on their system. But a
simple thread in StackOverflow gave me the solution to my problem. So
that's how it feels when you release an application that you created on
another system. Such a pleasure :) Just updating the version into 0.3.
Development for this version is now ongoing.