Coding is somehow stress-free than playing games XD
One major changes that I commit is the arrangement of classes. You can
now see a simple outline of the classes created on this file (see top).
Also, some minor changes on the names of the classes used. I find it
appealing when I only use one word when naming classes. Finally, I added
a new template for 'filing'. No widgets yet but somehow I was able to
connect it to the main window.