Added a QSpinBox for adding number of days under Search template
Done adding a new widget for search.py. The new widget is not
functioning yet, I was diverted doodling how to the fix layout to make
it pleasing to user. But more important is to keep on going and fix that
issue in the near future.

Things to do next_
[ ] FILING: Implement the whole logic of filing.py, you're running out
of time
[ ] SEARCHING: Implement the logic of self.daysSpinBox, this will take
so be prepared

[x] SEARCHING: add a dial in search.py wherein a user can simply enter
the number of days, wouldn't be that easy instead of hunting days on the
calendar? Just saying ~author # DONE - 2 Nov 2015 @ 1:43PM