Closes Issue #1 - new tracker feature
The new template tracker is now functioning as intended. It's now up to
the user on how they will respond to your application. I thought this
new feature is going to take time to implement but a simple list do the
amazing job for me. And you have still a day to spare before the 2-week
testing phase. Job well done :)