Initial commit for 2016
Happy New Year! The application has already been uploaded on the
company's server over a month ago and it's ready for the announcement.
I'm still working on the user's documentation before announcing it to
the whole team. I have another 2 months to start coding this app again
and I am looking forward for the upcoming bugs that will appear during
this coding timeframe.