Messing-up my repository
Remove the logging sessions after I recieved an error while running the application in .exe-mode. Also, that icon in the main window is not properly working after packaging the whole files. Googled everthing and I ended up with a messy code base full of comments! The whole code base are still functioning as intended, I am not just satisfied when the application is translated to executable. Anyways, just enjoy this release.