Early commit for v0.5
It seem's that I am too excited for the Betty's next patch version :) The development for 0.5 will official starts on the 1st of July but here I am sending my initial commit for the next version. Almost all of the files has been amended but the only visible enhancement can be seen in the main window title. Also added some TODO chores for this project.