Applied partial model/view paradigm to ISSUE #8
Somehow, I was able to implement model/view-esque on this project. The list in the new.py can now be edited. Next step would be to add an icon identifer on each list but the QModelIndex concept is too much for me to comprehend for now. Just need a break for this.