Piece of cake
I didn't see this coming. I somehow managed to re-implement the use of
the tracker. The tracker feature is now model/view implemented. Users
can now edit the entries inside the list by means of a double-clicked
and the main window will show the corresponding template by a simple
clicked on the one of entries. Pretty little neat but I still need to
add an icon identifier in the tracker.