Almost done with Issue #12
Today, I was able to get what I really want for this current issue.
Seeing that cool icons being added in the Tracker is very satisfying.
Next step would be to test this new implementation if it works to the
Team. I already froze an alpha version of v0.5. Testing will start
tomorrow.