Making it Pythonic so far
While waiting for my colleagues to somehow submit a report regarding 0.5
test, why not refine your unprofessional code just a bit?

I was able to refine some code snippets on the following files that was
commited. Remove some duplicates, refactor some variables and functions
and added a bit of docstrings on some methods.  This is another fun
coding session. I was able to appreciate the fact of having a good
readable code base. A professional coder knows how to write readable
codes! #lessonslearned