Filing form now resize as wanted
I was able to set the size of the special instruction field. It's just a
matter of tweaking the sizePolizy of this widget. Also, this widget is
now connected to the on_setCriteria_changed method. The text inputted in
this new field does show in the Preview field. However, the newline is
making my head bleed again. It is not displaying as intended. I guess I
need a break to think for the solution for this issue.