Nobody saves the day today
After tweaking the code that solved my problem yesterday, I reverted
back to the code that causes the same problem. It is weird because the
code that causes the crash now actually solves it. I have some
speculation perhaps it's the passing the SETTINGS_VERSION variable to
the restoreState() function or something else, I just don't know.
Anyway, these line of codes now solves my issue in #1. Also done a bit
of cleaning in this file.