Backporting to Python 3.5.2 from 3.6.0
After developing the Betty 0.7 for almost three (3) months in 3.6, I
ended up implementing the app in 3.5.2. This is due to the freezing
library that I use which is PyInstaller. This library doesn't yet
support Python 3.6! So the solution is to backport Python 3.5.2 as I am
not currently comfortable in using other freezing library (cx_freeze
py2exe, etc.). Anyway, this will be the final commit for 0.7 and now
ready of the Jun 1 release day.